fixed code review . added some widgets and some function fo loging page.
