+++ lib/common/styles/spacing_styles.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../utils/constants/sizes.dart';
+import '../../utils/constants/sizes.dart';
 
 class TSpacingStyle{
   static const EdgeInsetsGeometry paddingWithAppBarHeight = EdgeInsets.only(
