+++ lib/common/widgets.login_signup/social_buttons.dart
@@ -1,0 +1,47 @@
+
+
+import 'package:flutter/material.dart';
+
+import '../../utils/constants/colors.dart';
+import '../../utils/constants/image_strings.dart';
+import '../../utils/constants/sizes.dart';
+
+class TSocialButton extends StatelessWidget {
+  const TSocialButton({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Container(
+          decoration: BoxDecoration(
+              border: Border.all(color: TColors.grey),
+              borderRadius: BorderRadius.circular(100)),
+          child: IconButton(
+              onPressed: () {},
+              icon: Image(
+                  width: TSizes.iconMd,
+                  height: TSizes.iconMd,
+                  image: AssetImage(TImages.google))),
+        ),
+        SizedBox(
+          width: TSizes.spaceBtwItems,
+        ),
+        Container(
+          decoration: BoxDecoration(
+              border: Border.all(color: TColors.grey),
+              borderRadius: BorderRadius.circular(100)),
+          child: IconButton(
+              onPressed: () {},
+              icon: Image(
+                  width: TSizes.iconMd,
+                  height: TSizes.iconMd,
+                  image: AssetImage(TImages.facebook))),
+        ),
+      ],
+    );
+  }
+}
