+++ lib/features/authentication/screens/login/login.dart
@@ -2,13 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:grocify/features/authentication/screens/login/widgets/loging_form.dart';
 import 'package:grocify/features/authentication/screens/login/widgets/loging_header.dart';
-import 'package:grocify/utils/constants/colors.dart';
 import 'package:grocify/utils/constants/text_strings.dart';
-import 'package:grocify/utils/helpers/helper_functions.dart';
-
-
-import '../../../../common/spacing_styles.dart';
-import '../../../../utils/constants/image_strings.dart';
+import '../../../../common/styles/spacing_styles.dart';
+import '../../../../common/widgets.login_signup/form_divider.dart';
+import '../../../../common/widgets.login_signup/social_buttons.dart';
 import '../../../../utils/constants/sizes.dart';
 
 class LoginScreen extends StatelessWidget {
@@ -16,7 +13,7 @@ class LoginScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final dark = THelperFunctions.isDarkMode(context);
+
 
     return Scaffold(
       body: SingleChildScrollView(
@@ -25,16 +22,14 @@ class LoginScreen extends StatelessWidget {
           child: Column(
             children: [
               /// Logo, Title and SubTitle
-              TLogingHeader(dark: dark),
-              const SizedBox(
-                height: TSizes.sm,
-              ),
+              TLogingHeader(),
+              const SizedBox(height: TSizes.sm,),
 
               /// Form
               TLogingForm(),
 
               /// Divider
-              TFromDivider(dark: dark),
+              TFromDivider(dividerText: TTexts.orSignInWith.capitalize!),
               SizedBox(
                 height: TSizes.spaceBtwSections,
               ),
@@ -48,80 +43,11 @@ class LoginScreen extends StatelessWidget {
   }
 }
 
-class TSocialButton extends StatelessWidget {
-  const TSocialButton({
-    super.key,
-  });
 
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Container(
-          decoration: BoxDecoration(
-              border: Border.all(color: TColors.grey),
-              borderRadius: BorderRadius.circular(100)),
-          child: IconButton(
-              onPressed: () {},
-              icon: Image(
-                  width: TSizes.iconMd,
-                  height: TSizes.iconMd,
-                  image: AssetImage(TImages.google))),
-        ),
-        SizedBox(
-          width: TSizes.spaceBtwItems,
-        ),
-        Container(
-          decoration: BoxDecoration(
-              border: Border.all(color: TColors.grey),
-              borderRadius: BorderRadius.circular(100)),
-          child: IconButton(
-              onPressed: () {},
-              icon: Image(
-                  width: TSizes.iconMd,
-                  height: TSizes.iconMd,
-                  image: AssetImage(TImages.facebook))),
-        ),
-      ],
-    );
-  }
-}
-class TFromDivider extends StatelessWidget {
-  const TFromDivider({
-    super.key,
-    required this.dark,
-  });
 
-  final bool dark;
 
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Flexible(
-            child: Divider(
-          color: dark ? TColors.darkerGrey : TColors.grey,
-          thickness: 0.5,
-          indent: 60,
-          endIndent: 5,
-        )),
-        Text(
-          TTexts.orSignInWith.capitalize!,
-          style: Theme.of(context).textTheme.labelMedium,
-        ),
-        Flexible(
-            child: Divider(
-          color: dark ? TColors.darkerGrey : TColors.grey,
-          thickness: 0.5,
-          indent: 5,
-          endIndent: 60,
-        )),
-      ],
-    );
-  }
-}
+
+
