+++ lib/features/authentication/screens/login/widgets/loging_header.dart
@@ -1,21 +1,17 @@
-
-
 import 'package:flutter/material.dart';
-
 import '../../../../../utils/constants/image_strings.dart';
 import '../../../../../utils/constants/sizes.dart';
 import '../../../../../utils/constants/text_strings.dart';
+import '../../../../../utils/helpers/helper_functions.dart';
 
 class TLogingHeader extends StatelessWidget {
   const TLogingHeader({
-    super.key,
-    required this.dark,
-  });
+    super.key,});
 
-  final bool dark;
 
   @override
   Widget build(BuildContext context) {
+    final dark = THelperFunctions.isDarkMode(context);
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
