+++ lib/features/authentication/screens/onboarding/widgets/onboarding_page.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import '../../../../../utils/constants/sizes.dart';
 import '../../../../../utils/helpers/helper_functions.dart';
 
-class onBoardingPage extends StatelessWidget {
-  const onBoardingPage({
+class OnBoardingPage extends StatelessWidget {
+  const OnBoardingPage({
     super.key,
     required this.image,
     required this.title,
