+++ lib/utils/theme/custom_theme/chip_theme.dart
@@ -5,7 +5,7 @@ class TChipTheme {
   TChipTheme._();
 
   static ChipThemeData lightChipTheme = ChipThemeData(
-    disabledColor: Colors.grey.withOpacity(0.4),
+    disabledColor: Colors.grey.withValues(alpha: 0.4),
     labelStyle: const TextStyle(color: Colors.black),
     selectedColor: Colors.blue,
     padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 12),
