+++ lib/utils/theme/custom_theme/text_theme.dart
@@ -15,10 +15,10 @@ class TTextTheme{
 
     bodyLarge: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.w500, color: Colors.black),
     bodyMedium: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.normal, color: Colors.black),
-    bodySmall: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.w500, color: Colors.black.withOpacity(0.5)),
+    bodySmall: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.w500, color: Colors.black.withValues(alpha: 0.5)),
 
     labelLarge: const TextStyle().copyWith(fontSize: 12.0,fontWeight: FontWeight.normal,color: Colors.black),
-    labelMedium: const TextStyle().copyWith(fontSize: 12.0,fontWeight: FontWeight.normal,color: Colors.black.withOpacity(0.5)),
+    labelMedium: const TextStyle().copyWith(fontSize: 12.0,fontWeight: FontWeight.normal,color: Colors.black.withValues(alpha: 0.5)),
   );
 
   // Customizable Dark Text theme
@@ -33,9 +33,9 @@ class TTextTheme{
 
     bodyLarge: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.w500, color: Colors.white),
     bodyMedium: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.normal, color: Colors.white),
-    bodySmall: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.w500, color: Colors.white.withOpacity(0.5)),
+    bodySmall: const TextStyle().copyWith(fontSize: 14.0,fontWeight: FontWeight.w500, color: Colors.white.withValues(alpha: 0.5)),
 
     labelLarge: const TextStyle().copyWith(fontSize: 12.0,fontWeight: FontWeight.normal,color: Colors.white),
-    labelMedium: const TextStyle().copyWith(fontSize: 12.0,fontWeight: FontWeight.normal,color: Colors.white.withOpacity(0.5)),
+    labelMedium: const TextStyle().copyWith(fontSize: 12.0,fontWeight: FontWeight.normal,color: Colors.white.withValues(alpha: 0.5)),
   );
 }
